enh: cleanup runc cli and add new cli command for get runc id by container name
